+++ horizontal_ruler.rb
